refactor(entrypoint): reorganize the components
